pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "defectgen"
version = "0.1.0"
description = "Industrial defect image generation via diffusion / flow matching"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torchvision",
    "numpy",
    "einops",
    "omegaconf",
    "pyyaml",
    "tqdm",
    "wandb",
    "scipy",
    "Pillow",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-xdist", "ruff", "mypy"]
mcp = ["mcp>=1.0"]

[project.scripts]
defectgen-mcp = "src.mcp.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
